// public/node-framework/library/http/interfaces/crudSchema.interface.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface CrudSchema {
  createButton?: {
    enable?: boolean;
    label?: string;
  };
  form?: {
    createTitle?: string;
    updateTitle?: string;
  };
  fields: {
    type?: "choice" | "text" | "bool";
    name: string;
    label: string;
    items?: {
      [k: string]: unknown;
    };
    ignoreGrid?: boolean;
    ignoreForm?: boolean;
    form?: {
      [k: string]: unknown;
    };
  }[];
}